+++ tests/test_table.py
@@ -508,3 +508,9 @@ def test_weird_colspan():
 
 def test_caption_containing_piped_wikilink():
     assert Table('{|\n|+a [[b|c]]\n|}').caption == 'a [[b|c]]'
+
+
+def test_cell_header():
+    c = Table('{|\n!1!!style="color:red;"|2\n|}').cells(row=0, column=1)
+    c.value = '3'  # invalidates the match cache
+    assert c.value == '3'